Apply menu to tables
