--- clients/admin-ui/src/features/common/table/cells/constants.ts
@@ -1,1 +1,0 @@
-export const COLLAPSE_BUTTON_TEXT = "show less";
