+++ clients/admin-ui/src/features/common/table/cells/index.tsx
@@ -1,3 +1,2 @@
 export * from "./ListExpandableCell";
-export * from "./MenuHeaderCell";
 export * from "./TagExpandableCell";
